test(pending-outflows): add failing tests for the single-dialog refactor
Add tests/unit/pendingOutflowsSingleDialog.test.tsx. It proves the list
must render one print-check dialog, call useCheckBankAccounts once for
N rows, and hide the Print button with no check settings.

Rename tests/unit/PrintCheckButtonCategory.test.tsx to
PrintCheckDialogCategory.test.tsx. Drive the dialog through
PendingOutflowsList and the row button, not PrintCheckButton alone. Add
three cases for shared form state: a row switch clears the memo and
account, a reopen clears the memo, and a same-row refetch keeps a typed
memo.

Update pendingOutflowCardPrintCheckGate.test.tsx for the design's new
gate: drop the PrintCheckButton mock, pass onPrintCheck to
PendingOutflowCard, and add a case that hides the Print button with no
onPrintCheck.

Three tests fail today, each for the reason the design names: the
account hook still runs once per row, a background refetch still
clobbers a typed memo, and PendingOutflowCard still ignores
onPrintCheck. No source change in this commit.
