+++ tests/unit/pendingOutflowCardPrintCheckGate.test.tsx
@@ -3,8 +3,8 @@ import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
-// Capability gate under test (design §3.4): `PrintCheckButton` renders only
-// when `isResolved && hasCapability('edit:pending_outflows')`.
+// Print-button gate under test (design §5.6): the inline Print button renders
+// only when `onPrintCheck && isResolved && hasCapability('edit:pending_outflows')`.
 const hasCapabilityMock = vi.fn();
 let isResolvedMock = true;
 vi.mock('@/hooks/usePermissions', () => ({
@@ -14,21 +14,76 @@ vi.mock('@/hooks/usePermissions', () => ({
   }),
 }));
 
-// Isolate the render-site gate from PrintCheckButton's own hooks/dialog.
-vi.mock('@/components/pending-outflows/PrintCheckButton', () => ({
-  PrintCheckButton: () => <button data-testid="print-check-button">Print check</button>,
-}));
-
 vi.mock('@/contexts/RestaurantContext', () => ({
-  useRestaurantContext: () => ({ selectedRestaurant: { id: 'restaurant-1' } }),
+  useRestaurantContext: () => ({
+    selectedRestaurant: { id: 'restaurant-1', restaurant: { name: 'Test Restaurant' }, restaurant_id: 'restaurant-1' },
+  }),
 }));
 
 vi.mock('@/hooks/usePendingOutflows', () => ({
   usePendingOutflowMutations: () => ({
     voidPendingOutflow: { mutate: vi.fn() },
     deletePendingOutflow: { mutate: vi.fn() },
+    updatePendingOutflow: { mutateAsync: vi.fn() },
   }),
   usePendingOutflowMatches: () => ({ data: [] }),
+}));
+
+// PrintCheckButton is no longer mocked (design §5.6 removes it from
+// PendingOutflowCard's render path), but until that lands, PendingOutflowCard
+// still mounts the real PrintCheckButton behind the capability gate. Stub its
+// three hooks with valid data so the gate under test is `isResolved` /
+// `hasCapability` / `onPrintCheck`, not an unrelated "no check settings yet".
+vi.mock('@/hooks/useCheckSettings', () => ({
+  useCheckSettings: () => ({
+    settings: {
+      id: 'set-1',
+      restaurant_id: 'restaurant-1',
+      business_name: 'Test Restaurant LLC',
+      business_address_line1: '123 Main St',
+      business_address_line2: null,
+      business_city: 'Austin',
+      business_state: 'TX',
+      business_zip: '78701',
+      bank_name: null,
+      print_bank_info: false,
+      routing_number: null,
+      signature_url: null,
+    },
+  }),
+}));
+
+vi.mock('@/hooks/useCheckBankAccounts', () => ({
+  useCheckBankAccounts: () => ({
+    accounts: [{
+      id: 'acct-1',
+      account_name: 'Operating',
+      bank_name: 'First National',
+      next_check_number: 1001,
+      print_bank_info: false,
+      routing_number: null,
+      account_number_last4: null,
+      is_default: true,
+    }],
+    defaultAccount: {
+      id: 'acct-1',
+      account_name: 'Operating',
+      bank_name: 'First National',
+      next_check_number: 1001,
+      print_bank_info: false,
+      routing_number: null,
+      account_number_last4: null,
+      is_default: true,
+    },
+    claimCheckNumbers: { mutateAsync: vi.fn() },
+    fetchAccountSecrets: vi.fn(),
+  }),
+}));
+
+vi.mock('@/hooks/useCheckAuditLog', () => ({
+  useCheckAuditLog: () => ({
+    logCheckAction: { mutateAsync: vi.fn() },
+  }),
 }));
 
 // Unrelated to this gate; only mounted when the (unopened) manual-match dialog
@@ -61,48 +116,62 @@ const baseOutflow: PendingOutflow = {
   chart_account: null,
 };
 
-const renderCard = (outflow: PendingOutflow = baseOutflow) => {
+const renderCard = (
+  outflow: PendingOutflow = baseOutflow,
+  onPrintCheck?: (outflow: PendingOutflow) => void,
+) => {
   const queryClient = new QueryClient({
     defaultOptions: { queries: { retry: false } },
   });
   return render(
     <QueryClientProvider client={queryClient}>
-      <PendingOutflowCard outflow={outflow} />
+      <PendingOutflowCard outflow={outflow} onPrintCheck={onPrintCheck} />
     </QueryClientProvider>,
   );
 };
 
-describe('PendingOutflowCard — PrintCheckButton capability gate (design §3.4)', () => {
+const findPrintButton = () => screen.queryByRole('button', { name: /^Print check for Acme Produce$/i });
+
+describe('PendingOutflowCard — Print button capability gate (design §5.6)', () => {
   beforeEach(() => {
     hasCapabilityMock.mockReset();
     isResolvedMock = true;
   });
 
-  it('hides PrintCheckButton when the capability context has not resolved yet, even if capability would be granted', () => {
+  it('hides the Print button when the capability context has not resolved yet, even if capability would be granted', () => {
     isResolvedMock = false;
     hasCapabilityMock.mockReturnValue(true);
 
-    renderCard();
+    renderCard(baseOutflow, vi.fn());
 
-    expect(screen.queryByTestId('print-check-button')).not.toBeInTheDocument();
+    expect(findPrintButton()).not.toBeInTheDocument();
   });
 
-  it('hides PrintCheckButton once resolved when the user lacks edit:pending_outflows', () => {
+  it('hides the Print button once resolved when the user lacks edit:pending_outflows', () => {
     isResolvedMock = true;
     hasCapabilityMock.mockReturnValue(false);
 
-    renderCard();
+    renderCard(baseOutflow, vi.fn());
 
-    expect(screen.queryByTestId('print-check-button')).not.toBeInTheDocument();
+    expect(findPrintButton()).not.toBeInTheDocument();
     expect(hasCapabilityMock).toHaveBeenCalledWith('edit:pending_outflows');
   });
 
-  it('shows PrintCheckButton once resolved when the user has edit:pending_outflows', () => {
+  it('shows the Print button once resolved when the user has edit:pending_outflows', () => {
     isResolvedMock = true;
     hasCapabilityMock.mockReturnValue(true);
 
-    renderCard();
+    renderCard(baseOutflow, vi.fn());
 
-    expect(screen.getByTestId('print-check-button')).toBeInTheDocument();
+    expect(findPrintButton()).toBeInTheDocument();
+  });
+
+  it('hides the Print button when no onPrintCheck is passed, even with capability granted', () => {
+    isResolvedMock = true;
+    hasCapabilityMock.mockReturnValue(true);
+
+    renderCard(baseOutflow, undefined);
+
+    expect(findPrintButton()).not.toBeInTheDocument();
   });
 });
